Implement diagnosekoder API
Implements first working version of an api that make use of the icd10 registry list
from no.nav.syfo.sm.Diagnosekoder, with a few adjustments being made to the data before
its made available.

Currently it will only respond with diagnosekode data at /diagnosekoder when a 'query'
parameter is specified in the query parameters. Also allows for specifying a max number
of elements to be returned.